// orion-knowledge/web/admin/src/request/types.ts
export enum DomainNodeType {
  NodeTypeFolder = 1,
  NodeTypeDocument = 2,
}

/** @format int32 */
export enum DomainNodeStatus {
  /** 草稿 */
  NodeStatusUnreleased = 0,
  /** 更新未发布 */
  NodeStatusDraft = 1,
  /** 已发布 */
  NodeStatusPublished = 2,
}
// ...
export enum ConstsNodeRagInfoStatus {
  /** 等待处理 */
  NodeRagStatusPending = "PENDING",
  /** 正在进行处理（文本分割、向量化等） */
  NodeRagStatusRunning = "RUNNING",
  /** 处理失败 */
  NodeRagStatusFailed = "FAILED",
  /** 处理成功 */
  NodeRagStatusSucceeded = "SUCCEEDED",
  /** 重新索引中 */
  NodeRagStatusReindexing = "REINDEX",
}

export enum ConstsNodePermName {
  /** 导航内可见 */
  NodePermNameVisible = "visible",
  /** 可被访问 */
  NodePermNameVisitable = "visitable",
  /** 可被问答 */
  NodePermNameAnswerable = "answerable",
}

export enum ConstsNodeAccessPerm {
  /** 完全开放 */
  NodeAccessPermOpen = "open",
  /** 部分开放 */
  NodeAccessPermPartial = "partial",
  /** 完全禁止 */
  NodeAccessPermClosed = "closed",
}
// ...
export interface DomainBatchMoveReq {
  ids: string[];
  kb_id: string;
  parent_id?: string;
}
// ...
export interface DomainCreateNodeReq {
  content?: string;
  content_type?: string;
  emoji?: string;
  kb_id: string;
  name: string;
  nav_id: string;
  parent_id?: string;
  position?: number;
  summary?: string;
  type: 1 | 2;
}
// ...
export interface DomainMoveNodeReq {
  id: string;
  kb_id: string;
  next_id?: string;
  parent_id?: string;
  prev_id?: string;
}

export interface DomainNavDocConfig {
  nav_ids?: string[];
  title?: string;
}

export interface DomainNodeActionReq {
  action: "delete";
  ids: string[];
  kb_id: string;
}

export interface DomainNodeContentChunkSSE {
  emoji?: string;
  name?: string;
  node_id?: string;
  node_path_names?: string[];
  summary?: string;
}

export interface DomainNodeGroupDetail {
  auth_group_id?: number;
  auth_ids?: number[];
  kb_id?: string;
  name?: string;
  node_id?: string;
  perm?: ConstsNodePermName;
}

export interface DomainNodeListItemResp {
  content_type?: string;
  created_at?: string;
  creator?: string;
  creator_id?: string;
  editor?: string;
  editor_id?: string;
  emoji?: string;
  id?: string;
  name?: string;
  nav_id?: string;
  parent_id?: string;
  permissions?: DomainNodePermissions;
  position?: number;
  publisher_id?: string;
  rag_info?: DomainRagInfo;
  status?: DomainNodeStatus;
  summary?: string;
  type?: DomainNodeType;
  updated_at?: string;
}

export interface DomainNodeMeta {
  content_type?: string;
  emoji?: string;
  summary?: string;
}

export interface DomainNodePermissions {
  /** 可被问答 */
  answerable?: ConstsNodeAccessPerm;
  /** 导航内可见 */
  visible?: ConstsNodeAccessPerm;
  /** 可被访问 */
  visitable?: ConstsNodeAccessPerm;
}

export interface DomainNodeSummaryReq {
  ids: string[];
  kb_id: string;
}
// ...
export interface DomainPWResponse {
  code?: number;
  data?: unknown;
  message?: string;
  success?: boolean;
}

export interface DomainPaginatedResultArrayDomainConversationMessageListItem {
  data?: DomainConversationMessageListItem[];
  total?: number;
}

export interface DomainProviderModelListItem {
  model?: string;
}

export interface DomainQuestionConfig {
  list?: {
    id?: string;
    question?: string;
  }[];
  title?: string;
  type?: string;
}

export interface DomainRagInfo {
  message?: string;
  status?: ConstsNodeRagInfoStatus;
  synced_at?: string;
}

export interface DomainRecommendNodeListResp {
  emoji?: string;
  id?: string;
  name?: string;
  nav_id?: string;
  nav_name?: string;
  parent_id?: string;
  permissions?: DomainNodePermissions;
  position?: number;
  recommend_nodes?: DomainRecommendNodeListResp[];
  summary?: string;
  type?: DomainNodeType;
}

export interface DomainResponse {
  data?: unknown;
  message?: string;
  success?: boolean;
}
// ...
export interface DomainUpdateNodeReq {
  content?: string;
  content_type?: string;
  emoji?: string;
  id: string;
  kb_id: string;
  name?: string;
  nav_id?: string;
  position?: number;
  summary?: string;
}
// ...
export interface GithubComOrionPlatformOrionKnowledgeApiNodeV1NodeListGroupNavResp {
  count?: number;
  is_released?: boolean;
  list?: DomainNodeListItemResp[];
  nav_id?: string;
  nav_name?: string;
  position?: number;
}
// ...
export interface V1NodeDetailResp {
  content?: string;
  created_at?: string;
  creator_account?: string;
  creator_id?: string;
  editor_account?: string;
  editor_id?: string;
  id?: string;
  kb_id?: string;
  meta?: DomainNodeMeta;
  name?: string;
  nav_id?: string;
  parent_id?: string;
  permissions?: DomainNodePermissions;
  publisher_account?: string;
  publisher_id?: string;
  pv?: number;
  status?: DomainNodeStatus;
  type?: DomainNodeType;
  updated_at?: string;
}

export interface V1NodeMoveNavReq {
  /** @minItems 1 */
  ids: string[];
  kb_id: string;
  nav_id: string;
}
// ...
export interface V1NodeRestudyReq {
  kb_id: string;
  /** @minItems 1 */
  node_ids: string[];
}

export type V1NodeRestudyResp = Record<string, any>;

export interface V1NodeStatsResp {
  /** 未发布的文档数 */
  unpublished_count?: number;
  /** 未发布目录数量 */
  unreleased_nav_count?: number;
  /** 未学习的文档数 */
  unstudied_count?: number;
}
// ...
export interface GetApiV1NodeDetailParams {
  format?: string;
  id: string;
  kb_id: string;
}

export interface GetApiV1NodeListParams {
  kb_id: string;
  nav_id?: string;
  search?: string;
}

export interface GetApiV1NodeListGroupNavParams {
  kb_id: string;
  nav_ids?: string[];
  search?: string;
  status?: "released" | "unpublished" | "unstudied";
}

export interface GetApiV1NodePermissionParams {
  id: string;
  kb_id: string;
}

export interface GetApiV1NodeRecommendNodesParams {
  kb_id: string;
  nav_ids?: string[];
  node_ids?: string[];
}

export interface GetApiV1NodeStatsParams {
  kb_id: string;
}